Create usersHandlers.js in the handlers folder to refactor my code
